UTCL-MFG-DATA_FOUNDATION_PLATEFORM/CBM-Functions/blaine_tag_master_update/requirements.txt
pyodbc
pandas
numpy
boto3
python-dotenv
requests
